KN-22930 review issues
