+++ Knoema.Client/Meta/DateRange.cs
@@ -8,11 +8,11 @@ namespace Knoema.Meta
 		public int Calendar { get; set; }
 		public DateTime StartDate { get; set; }
 		public DateTime EndDate { get; set; }
-		public List<char> Frequencies { get; set; }
+		public List<string> Frequencies { get; set; }
 
 		public DateRange()
 		{
-			Frequencies = new List<char>();
+			Frequencies = new List<string>();
 		}
 	}
 }
